+++ CHANGELOG.md
@@ -8,15 +8,22 @@
 
 - Use Hierarchy Builder to define the ordType interface.
 
+- `ordMixin` has been replaced with `hasOrd`
+
+- Use maximally implicit arguments for the type arguments of `getm`, `setm`,
+  `repm`, `updm`, `mapim`, `mapm`, `filterm`, `remm`, `mkfmap`, `mkfmapf`,
+  `mkfmapfp` and `domm`.
+
+## Deprecated
+
 - `InjOrdMixin`, `PcanOrdMixin` and `CanOrdMixin` have been deprecated in favor
   of `InjHasOrd`, `PcanHasOrd` and `CanHasOrd`.
 
 - The `[ordMixin of T by <:]` notation has been deprecated in favor of `[Ord of
   T by <:]`.
 
-- Use maximally implicit arguments for the type arguments of `getm`, `setm`,
-  `repm`, `updm`, `mapim`, `mapm`, `filterm`, `remm`, `mkfmap`, `mkfmapf`,
-  `mkfmapfp` and `domm`.
+- The `[derive ordMixin for T]` notation has been deprecated in favor of
+  `[derive hasOrd for T]`.
 
 ## Removed